bugfix: interruption when editing operator properties
when typing in operator property editor, the selected operator gets un-highlighted, closing the property editor.

This was due to improper use of Array.splice.
Normal behavior is to unhighlight all other operators except the one being edited. However, the operator being edited was not properly spliced out of the list to be unhighlighed